bug solved

+++ tree/Node.java
@@ -1,10 +1,10 @@
 package tree;
 
 public class Node {
-    public int data ;
+    public Double data ;
     public Node left , right ;
 
-    public Node(int data){
+    public Node(Double data){
         this.data = data ;
         left = right = null ;
     }
